Subtract from char to get int
When calculating the value of the card, iterate through the hand vector.
If the rank is an Ace, add 11 to the total, else for the other named
cards, add 10.

For the numbered cards, subtract the char '0' (zero) from the rank to
get the int value of the card and add it to the total hand value.

NOTE: This does not work if the char is greater than 10.